+++ discord/events/__init__.py
@@ -21,6 +21,9 @@ from .guild_scheduled_event_user_remove import GuildScheduledEventUserRemoveEven
 from .guild_soundboard_sound_delete import GuildSoundboardSoundDeleteEvent
 from .guild_soundboard_sounds_update import GuildSoundboardSoundsUpdateEvent
 from .guild_stickers_update import GuildStickersUpdateEvent
+from .integration_create import IntegrationCreateEvent
+from .integration_delete import IntegrationDeleteEvent
+from .integration_update import IntegrationUpdateEvent
 from .presence_update import PresenceUpdateEvent
 from .ready import ReadyEvent
 from .soundboard_sounds import SoundboardSoundsEvent