+++ discord/events/integration_update.py
@@ -1,0 +1,7 @@
+from ..objects import Integration
+from ..snowflake import Snowflake
+
+
+class IntegrationUpdateEvent(Integration):
+  guild_id: Snowflake
+  """ID of the guild."""